# swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.2"
  title: Dynamo Auto Categorization
  description: Categorization API  
securityDefinitions:
 authorization:
  description: |
   For accessing the API a valid JWT token must be passed in all the queries in
   the 'Authorization' header.
   A valid JWT token is generated by the API and retourned as answer of a call
   to the route /login giving a valid user & password.
   The following syntax must be used in the 'Authorization' header :
       Bearer: xxxxxx.yyyyyyy.zzzzzz
  type: apiKey
  name: authorization
  in: header
# during dev, should point to your local machine
host: localhost:3030
# basePath prefixes all resource paths
basePath: /
#
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
paths:
  /swagger_test:
    x-swagger-router-controller: cat_controller
    get:
      tags:
        - public
      description: Test swagger
      operationId: test
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /create:
    x-swagger-router-controller: cat_controller
    post:
      security:
        - authorization: []
      tags:
        - public
      description: Create initial tenant configuration 
      operationId: create
      parameters:
        - name: data
          schema:
            $ref: "#/definitions/TenantData"
          description: data necesseary to initialize categorization for tenant
          in: body
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /configure_domain:
    x-swagger-router-controller: cat_controller
    post:
      security:
        - authorization: []
      tags:
      - public
      description: maps specific word to specific categorry and increases the confidence with provided threshold
      operationId: configure_domain
      parameters:
        - name: data
          schema:
            $ref: "#/definitions/DomainData"
          description: data necesseary to execute
          in: body
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /finalize:
    x-swagger-router-controller: cat_controller
    delete:
      security:
        - authorization: []
      tags:
      - public
      description: Delete tenant configuration
      operationId: finalize_tenant
      parameters:
        - name: data
          schema:
            $ref: "#/definitions/TenantData"
          in: body
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /add_training_data:
    x-swagger-router-controller: cat_controller
    post:
      security:
        - authorization: []
      tags:
      - public
      description: Add simple training data in categorization nomenclature. Data consists of document title only
      operationId: add_training_data
      parameters:
        - name: data
          schema:
            $ref: "#/definitions/DocumentData"
          description: data necesseary to execute
          in: body
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /add_training_data_attachment:
    x-swagger-router-controller: cat_controller
    post:
      security:
        - authorization: []
      tags:
      - public
      description: Add training data for document with its attachment
      operationId: add_training_data_attachment
      consumes:
        - multipart/form-data
      parameters:
        - in: formData
          name: attachment
          description: The attachment to upload
          type: file    
        - in: formData
          name: unique_id
          type: string    
        - in: formData
          name: domain
          type: string    
        - in: formData
          name: document_id
          type: string    
        - in: formData
          name: title
          type: string    
        - in: formData
          name: ignore_title
          type: boolean
        - in: formData
          name: word_threshold
          type: integer
        - in: formData
          name: categories
          type: array
          items:
            type: string    
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /add_training_data_email:
    x-swagger-router-controller: cat_controller
    post:
      security:
        - authorization: []
      tags:
      - public
      description: Add training data for emails
      consumes:
        - multipart/form-data
      parameters:
        - in: formData
          name: attachment
          description: The attachment to upload
          type: file    
        - in: formData
          name: unique_id
          type: string    
        - in: formData
          name: domain
          type: string    
        - in: formData
          name: document_id
          type: string    
        - in: formData
          name: email_subject
          type: string    
        - in: formData
          name: email_body
          type: string    
        - in: formData
          name: email_to
          type: array
          items:
            type: string    
        - in: formData
          name: email_from
          type: string    
        - in: formData
          name: word_threshold
          type: integer
        - in: formData
          name: categories
          type: array
          items:
            type: string    
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /train:
    x-swagger-router-controller: cat_controller
    post:
      security:
        - authorization: []
      tags:
      - public
      description: force train new categorization data
      operationId: train
      parameters:
        - name: data
          schema:
            $ref: "#/definitions/TrainingData"
          description: data necesseary to execute
          in: body
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /get_training_status:
    x-swagger-router-controller: cat_controller
    get:
      security:
        - authorization: []
      tags:
        - public
      description: Check status of training
      operationId: get_training_status
      parameters:
        - name: unique_id
          in: query
          type: string
          required: true
        - name: domain
          in: query
          type: string
          required: true
      responses:
        "200":
          description: Success
          schema:
            type: string
            description: return empty/in_progress/completed/failed
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /predict:
    x-swagger-router-controller: cat_controller
    get:
      security:
        - authorization: []
      tags:
        - public
      description: Check status of training
      operationId: predict
      parameters:
        - name: unique_id
          in: query
          type: string
          required: true
        - name: domain
          in: query
          type: string
          required: true
        - name: title
          in: query
          type: string
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/PredictionResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /predict_document:
    x-swagger-router-controller: cat_controller
    post:
      security:
        - authorization: []
      tags:
      - public
      description: Add training data for document with its attachment
      operationId: predict_document
      consumes:
        - multipart/form-data
      parameters:
        - in: formData
          name: attachment
          description: The attachment to upload
          type: file    
        - in: formData
          name: unique_id
          type: string    
        - in: formData
          name: domain
          type: string    
        - in: formData
          name: document_id
          type: string    
        - in: formData
          name: title
          type: string    
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/PredictionResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /predict_email:
    x-swagger-router-controller: cat_controller
    post:
      security:
        - authorization: []
      tags:
      - public
      description: Add training data for document with its attachment
      operationId: predict_email
      consumes:
        - multipart/form-data
      parameters:
        - in: formData
          name: attachment
          description: The attachment to upload
          type: file    
        - in: formData
          name: unique_id
          type: string    
        - in: formData
          name: domain
          type: string    
        - in: formData
          name: document_id
          type: string    
        - in: formData
          name: email_subject
          type: string    
        - in: formData
          name: email_body
          type: string    
        - in: formData
          name: email_to
          type: array
          items:
            type: string    
        - in: formData
          name: email_from
          type: string    
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/PredictionResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /correct:
    x-swagger-router-controller: cat_controller
    post:
      security:
        - authorization: []
      tags:
      - public
      description: supplies wright categories when prediction is not correct
      operationId: correct_model
      parameters:
        - name: data
          schema:
            $ref: "#/definitions/DocumentData"
          description: data necesseary to execute
          in: body
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /confusion_matrix:
    x-swagger-router-controller: cat_controller
    get:
      security:
        - authorization: []
      tags:
      - public
      description: Creates confusion matrix for error results anlytics
      operationId: confusion_matrix
      parameters:
        - name: unique_id
          in: query
          type: string
          required: true
        - name: domain
          in: query
          type: string
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ConfusionMatrixResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /evaluate:
    x-swagger-router-controller: cat_controller
    get:
      security:
        - authorization: []
      tags:
      - public
      description: Get prediction score
      operationId: evaluate
      parameters:
        - name: unique_id
          in: query
          type: string
          required: true
        - name: domain
          in: query
          type: string
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /alter_boost:
    x-swagger-router-controller: cat_controller
    post:
      tags:
      - public
      description: maps specific word to specific categorry and increases the confidence with provided threshold
      operationId: alter_boost
      parameters:
        - name: data
          schema:
            $ref: "#/definitions/BoostData"
          description: data necesseary to execute
          in: body
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/SuccessResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
definitions:
  TenantData:
    type: object
    properties:
      unique_id:
        type: string
        description: tenant unique identification, can be tenant id, tenant name, combination from tenant name + slot id, etc.
      human_readable:
        type: string
        description: human readable tenant name, for visualization in graphics
  BoostData:
    type: object
    properties:
      unique_id:
        type: string
      domain:
        type: string
        description: domain give possibilities to train several models for any tenant, can be used for categorizing similar emtities like activity for example
      boosts:
        type: array
        items:
          type: object
          description: array of objects containing ( term, category to be boosted, min_threshold and boost number). in object clear property removes an existing boost
  DocumentData:
    type: object
    properties:
      unique_id:
        type: string
      domain:
        type: string
        description: domain give possibilities to train several models for any tenant, can be used for categorizing similar emtities like activity for example
      title:
        type: string
      categories:
        type: array
        items:
          type: string
  TrainingData:
    type: object
    properties:
      unique_id:
        type: string
      domain:
        type: string
  DomainData:
    type: object
    properties:
      tenant_uid:
        type: string
      domain:
        type: string
        description: domain give possibilities to train several models for any tenant, can be used for categorizing similar emtities like activity for example
      options:
        type: object
      boosts:
        type: object
  SuccessResponse:
    type: object
    properties:
      text:
        type: string
  ErrorResponse:
    required:
    - code
    - text
    properties:
      code:
        type: integer
      text:
        type: string
  PredictionResponse:
    properties:
      result:
        type: object
        properties:
          tags:
            type: array
            items:
              type: string
      error:
        $ref: "#/definitions/ErrorResponse"
  ConfusionMatrixResponse:
    properties:
      result:
        type: object
        properties:
          matrix:
            type: array
            items:
              type: array
              items: 
                type: integer
          categories:
            type: array
            items:
              type: string
      error:
        $ref: "#/definitions/ErrorResponse"